day1

+++ Type_casting.java
@@ -1,8 +1,8 @@
 public class Casting  {
     public static void main(String[] args)  {
-        int x = 7;  
-        long y = x;  
-        float z = y;  
+        double x = 7;  
+        float y = x;  
+        int z = y;  
         System.out.println("Before conversion, int value "+x);  
         System.out.println("After conversion, long value "+y);  
         System.out.println("After conversion, float value "+z);
